+++ src/uart.c
@@ -338,7 +338,7 @@ bool uart_out_string(uint8_t uart_channel, char* data, uint8_t size)
     bool status = true;
 
     // Write one byte at a time
-    uint8_t i = 0;
+    int i = 0;
     for (i = 0; (i < size) && (status) && (data[i] != '\0'); i++)
     {
         status &= uart_out_byte(uart_channel, (uint8_t) data[i]);
